test: split 6 more files (round 3 of flake reduction)
Continued splitting of files >1000 lines to further reduce per-file
test count and limit blast radius of any single timeout cascade.

  starting-screens-write.test.js (1201)         → -validation, -behavior
  starting-screens-read-extras.test.js (1196)   → -a, -b
  admin-users-missing-core.test.js (1188)       → -misc, -warnings, -auth
  reports-coverage-reports-core.test.js (1110)  → -a, -b
  identity-graph-write.test.js (1070)           → -lifecycle, -admin
  admin-suggestions-new-routes-core.test.js (1066) → -write, -read

Largest remaining file: 1444 lines.
