+++ express-api/tests/routes/admin-users-missing-core-misc.test.js
@@ -167,318 +167,17 @@ describe('GET /api/user/:uniqueId', () => {
 
 // ─── POST /api/user/:uniqueId/warn ───────────────────────────────
 
-describe('POST /api/user/:uniqueId/warn', () => {
-  it('returns 403 for non-admin', async () => {
-    blockAdmin();
-    const app = createAdminApp();
-    const res = await request(app)
-      .post('/api/user/10000001/warn')
-      .send({ reason: 'Spamming', severity: 2 });
-    expect(res.status).toBe(403);
-  });
-
-  it('returns 400 when reason is missing', async () => {
-    const app = createAdminApp();
-    const res = await request(app).post('/api/user/10000001/warn').send({ severity: 2 });
-    expect(res.status).toBe(400);
-    expect(res.body.error).toMatch(/reason/i);
-  });
-
-  it('returns 400 when severity is out of range', async () => {
-    const app = createAdminApp();
-    const res = await request(app)
-      .post('/api/user/10000001/warn')
-      .send({ reason: 'Bad behaviour', severity: 6 });
-    expect(res.status).toBe(400);
-    expect(res.body.error).toMatch(/severity/i);
-  });
-
-  it('returns 404 when target user does not exist', async () => {
-    // createWarning: first reads db.doc().get() for the user — not found → throws before getDoc
-    mockDocGet.mockResolvedValueOnce({ exists: false });
-
-    const app = createAdminApp();
-    const res = await request(app)
-      .post('/api/user/10000001/warn')
-      .send({ reason: 'Spamming', severity: 2 });
-    expect(res.status).toBe(404);
-  });
-
-  it('returns 200 with warningId, newGcs, deduction on success', async () => {
-    // createWarning: 1st db.doc().get() = user doc, 2nd getDoc = admin doc
-    mockDocGet.mockResolvedValueOnce({
-      exists: true,
-      id: '10000001',
-      data: () => ({
-        gcsScore: 90,
-        warningCount: 1,
-      }),
-    });
-    getDoc.mockResolvedValueOnce({ displayName: 'Admin User' }); // admin user lookup
-
-    const app = createAdminApp();
-    const res = await request(app)
-      .post('/api/user/10000001/warn')
-      .send({ reason: 'Inappropriate content', severity: 2 });
-
-    expect(res.status).toBe(200);
-    expect(res.body.success).toBe(true);
-    expect(res.body.warningId).toBeDefined();
-    expect(res.body.newGcs).toBe(80); // 90 - 10 (severity 2 deduction)
-    expect(res.body.deduction).toBe(10);
-    expect(res.body.warningCount).toBe(2);
-  });
-});
-
 // ─── GET /api/user/:uniqueId/warnings ────────────────────────────
 
-describe('GET /api/user/:uniqueId/warnings', () => {
-  it('returns 403 for non-admin', async () => {
-    blockAdmin();
-    const app = createAdminApp();
-    const res = await request(app).get('/api/user/10000001/warnings');
-    expect(res.status).toBe(403);
-  });
-
-  it('returns 200 with empty warnings list when none exist', async () => {
-    mockCollectionGet = jest.fn().mockResolvedValue({ docs: [] });
-    const app = createAdminApp();
-    const res = await request(app).get('/api/user/10000001/warnings');
-    expect(res.status).toBe(200);
-    expect(res.body.warnings).toEqual([]);
-    expect(res.body.hasMore).toBe(false);
-  });
-
-  it('returns 200 with warnings and hasMore=false when under limit', async () => {
-    const warnDocs = [
-      {
-        id: 'warn-1',
-        data: () => ({ reason: 'Spam', severity: 1, createdAt: 1709913600000, revoked: false }),
-      },
-      {
-        id: 'warn-2',
-        data: () => ({
-          reason: 'Harassment',
-          severity: 3,
-          createdAt: 1709913500000,
-          revoked: false,
-        }),
-      },
-    ];
-    mockCollectionGet = jest.fn().mockResolvedValue({ docs: warnDocs });
-
-    const app = createAdminApp();
-    const res = await request(app).get('/api/user/10000001/warnings?limit=20');
-    expect(res.status).toBe(200);
-    expect(res.body.warnings).toHaveLength(2);
-    expect(res.body.hasMore).toBe(false);
-  });
-
-  it('returns hasMore=true when there are more results than the limit', async () => {
-    // Route fetches limit+1 docs to detect "has more"
-    const warnDocs = Array.from({ length: 4 }, (_, i) => ({
-      id: `warn-${i}`,
-      data: () => ({ reason: 'Test', severity: 1, createdAt: 1709913600000 - i }),
-    }));
-    mockCollectionGet = jest.fn().mockResolvedValue({ docs: warnDocs });
-
-    const app = createAdminApp();
-    // Request limit=3, we return 4 docs → hasMore should be true
-    const res = await request(app).get('/api/user/10000001/warnings?limit=3');
-    expect(res.status).toBe(200);
-    expect(res.body.warnings).toHaveLength(3);
-    expect(res.body.hasMore).toBe(true);
-  });
-});
-
 // ─── POST /api/user/:id/warnings/:warnId/revoke ──────────────────
-
-describe('POST /api/user/:uniqueId/warnings/:warningId/revoke', () => {
-  it('returns 403 for non-admin', async () => {
-    blockAdmin();
-    const app = createAdminApp();
-    const res = await request(app).post('/api/user/10000001/warnings/warn-1/revoke');
-    expect(res.status).toBe(403);
-  });
-
-  it('returns 404 when warning does not exist', async () => {
-    getDoc.mockResolvedValueOnce(null); // warning doc not found
-
-    const app = createAdminApp();
-    const res = await request(app).post('/api/user/10000001/warnings/warn-999/revoke');
-    expect(res.status).toBe(404);
-    expect(res.body.error).toMatch(/warning not found/i);
-  });
-
-  it('returns 400 when warning is already revoked', async () => {
-    getDoc.mockResolvedValueOnce({
-      id: 'warn-1',
-      revoked: true,
-      gcsDeduction: 10,
-    });
-
-    const app = createAdminApp();
-    const res = await request(app).post('/api/user/10000001/warnings/warn-1/revoke');
-    expect(res.status).toBe(400);
-    expect(res.body.error).toMatch(/already revoked/i);
-  });
-
-  it('returns 200 and restores GCS on successful revoke', async () => {
-    // First getDoc call = warning doc
-    getDoc.mockResolvedValueOnce({
-      id: 'warn-1',
-      revoked: false,
-      gcsDeduction: 10,
-    });
-    // Second db.doc().get() call = user doc (for current GCS)
-    mockDocGet.mockResolvedValueOnce({
-      exists: true,
-      id: '10000001',
-      data: () => ({ gcsScore: 80, warningCount: 2 }),
-    });
-
-    const app = createAdminApp();
-    const res = await request(app).post('/api/user/10000001/warnings/warn-1/revoke');
-
-    expect(res.status).toBe(200);
-    expect(res.body.success).toBe(true);
-    expect(res.body.restoredGcs).toBe(90); // 80 + 10
-    expect(res.body.deduction).toBe(10);
-
-    // Warning should be marked revoked
-    expect(mockDocUpdate).toHaveBeenCalledWith(
-      'users/10000001/warnings/warn-1',
-      expect.objectContaining({ revoked: true, revokedBy: 'admin-uid' }),
-    );
-    // User GCS should be restored
-    expect(mockDocUpdate).toHaveBeenCalledWith(
-      'users/10000001',
-      expect.objectContaining({ gcsScore: 90, warningCount: 1 }),
-    );
-  });
-});
 
 // ─── POST /api/user/:uniqueId/reset-gcs ──────────────────────────
 
 // ─── GET /api/user/:uniqueId/stalkers ────────────────────────────
 
-describe('GET /api/user/:uniqueId/stalkers', () => {
-  it('returns 403 for non-admin', async () => {
-    blockAdmin();
-    const app = createAdminApp();
-    const res = await request(app).get('/api/user/10000001/stalkers');
-    expect(res.status).toBe(403);
-  });
-
-  it('returns 200 with empty stalkers array when no stalkers exist', async () => {
-    mockCollectionGet = jest.fn().mockResolvedValue({ docs: [] });
-    const app = createAdminApp();
-    const res = await request(app).get('/api/user/10000001/stalkers');
-    expect(res.status).toBe(200);
-    expect(res.body.stalkers).toEqual([]);
-    expect(res.body.count).toBe(0);
-  });
-
-  it('returns 200 with stalker IDs and count when stalkers exist', async () => {
-    mockCollectionGet = jest.fn().mockResolvedValue({
-      docs: [{ id: 'stalker-A' }, { id: 'stalker-B' }, { id: 'stalker-C' }],
-    });
-
-    const app = createAdminApp();
-    const res = await request(app).get('/api/user/10000001/stalkers');
-    expect(res.status).toBe(200);
-    expect(res.body.stalkers).toEqual(['stalker-A', 'stalker-B', 'stalker-C']);
-    expect(res.body.count).toBe(3);
-  });
-});
-
 // ─── GET /api/user/:uniqueId — normalizeUser suspended user branches ──
 
 // ─── GET /api/user/:uniqueId — backfillAuthInfo branch ──
-
-describe('GET /api/user/:uniqueId — backfillAuthInfo', () => {
-  it('backfills email from Firebase Auth when user doc has no email', async () => {
-    const { auth } = require('../../src/utils/firebase');
-    auth.getUser.mockResolvedValueOnce({
-      uid: 'uid-no-email',
-      email: 'backfilled@example.com',
-      providerData: [],
-    });
-
-    mockDocGet.mockResolvedValueOnce({
-      exists: true,
-      id: '10000003',
-      data: () => ({
-        uniqueId: 10000003,
-        displayName: 'NoEmail',
-        gcsScore: 100,
-        gcsLastDeductionAt: null,
-        email: null,
-        firebaseUid: 'uid-no-email',
-      }),
-    });
-
-    const app = createAdminApp();
-    const res = await request(app).get('/api/user/10000003');
-    expect(res.status).toBe(200);
-    expect(res.body.email).toBe('backfilled@example.com');
-    // The backfill should also update the Firestore doc (fire-and-forget)
-    expect(mockDocUpdate).toHaveBeenCalledWith(
-      'users/10000003',
-      expect.objectContaining({ email: 'backfilled@example.com' }),
-    );
-  });
-
-  it('falls back to provider email when Firebase Auth has no direct email', async () => {
-    const { auth } = require('../../src/utils/firebase');
-    auth.getUser.mockResolvedValueOnce({
-      uid: 'uid-provider',
-      email: null,
-      providerData: [{ providerId: 'google.com', email: 'provider@gmail.com' }],
-    });
-
-    mockDocGet.mockResolvedValueOnce({
-      exists: true,
-      id: '10000004',
-      data: () => ({
-        uniqueId: 10000004,
-        displayName: 'ProviderOnly',
-        gcsScore: 100,
-        gcsLastDeductionAt: null,
-        email: null,
-        firebaseUid: 'uid-provider',
-      }),
-    });
-
-    const app = createAdminApp();
-    const res = await request(app).get('/api/user/10000004');
-    expect(res.status).toBe(200);
-    expect(res.body.email).toBe('provider@gmail.com');
-  });
-
-  it('handles Firebase Auth lookup failure gracefully', async () => {
-    const { auth } = require('../../src/utils/firebase');
-    auth.getUser.mockRejectedValueOnce(new Error('Auth service down'));
-
-    mockDocGet.mockResolvedValueOnce({
-      exists: true,
-      id: '10000005',
-      data: () => ({
-        uniqueId: 10000005,
-        displayName: 'AuthFail',
-        gcsScore: 100,
-        gcsLastDeductionAt: null,
-        email: null,
-        firebaseUid: 'uid-auth-fail',
-      }),
-    });
-
-    const app = createAdminApp();
-    const res = await request(app).get('/api/user/10000005');
-    expect(res.status).toBe(200);
-    expect(res.body.email).toBeNull();
-  });
-});
 
 // ─── GET /api/user/:uniqueId — 500 error branch ──
 
@@ -494,45 +193,6 @@ describe('GET /api/user/:uniqueId — error handling', () => {
 });
 
 // ─── GET /api/user/:uid/auth-debug ──────────────────────────────────
-
-describe('GET /api/user/:uid/auth-debug', () => {
-  it('returns 403 for non-admin', async () => {
-    blockAdmin();
-    const app = createAdminApp();
-    const res = await request(app).get('/api/user/firebase-uid-1/auth-debug');
-    expect(res.status).toBe(403);
-  });
-
-  it('returns 200 with raw Firebase Auth data for admin', async () => {
-    const { auth } = require('../../src/utils/firebase');
-    auth.getUser.mockResolvedValueOnce({
-      uid: 'firebase-uid-1',
-      email: 'user@example.com',
-      providerData: [{ providerId: 'google.com' }],
-      disabled: false,
-      metadata: { creationTime: '2024-01-01' },
-    });
-
-    const app = createAdminApp();
-    const res = await request(app).get('/api/user/firebase-uid-1/auth-debug');
-    expect(res.status).toBe(200);
-    expect(res.body.uid).toBe('firebase-uid-1');
-    expect(res.body.email).toBe('user@example.com');
-    expect(res.body.disabled).toBe(false);
-    expect(res.body.providerData).toEqual([{ providerId: 'google.com' }]);
-    expect(res.body.metadata).toBeDefined();
-  });
-
-  it('returns 500 when Firebase Auth lookup throws', async () => {
-    const { auth } = require('../../src/utils/firebase');
-    auth.getUser.mockRejectedValueOnce(new Error('Auth service unavailable'));
-
-    const app = createAdminApp();
-    const res = await request(app).get('/api/user/bad-uid/auth-debug');
-    expect(res.status).toBe(500);
-    expect(res.body.error).toMatch(/internal server error/i);
-  });
-});
 
 // ─── PATCH /api/user/:uniqueId — additional branches ────────────────
 
@@ -585,178 +245,15 @@ describe('PATCH /api/user/:uniqueId — additional branches', () => {
 
 // ─── POST /api/user/:uniqueId/notify-changes ────────────────────────
 
-describe('POST /api/user/:uniqueId/notify-changes', () => {
-  it('returns 403 for non-admin', async () => {
-    blockAdmin();
-    const app = createAdminApp();
-    const res = await request(app)
-      .post('/api/user/10000001/notify-changes')
-      .send({ fields: ['displayName'] });
-    expect(res.status).toBe(403);
-  });
-
-  it('returns 400 when fields is not a non-empty array', async () => {
-    const app = createAdminApp();
-    const res = await request(app).post('/api/user/10000001/notify-changes').send({ fields: [] });
-    expect(res.status).toBe(400);
-    expect(res.body.error).toMatch(/fields.*non-empty/i);
-  });
-
-  it('returns 400 when fields is missing', async () => {
-    const app = createAdminApp();
-    const res = await request(app).post('/api/user/10000001/notify-changes').send({});
-    expect(res.status).toBe(400);
-  });
-
-  it('returns notified=false when fields are not notifiable', async () => {
-    const app = createAdminApp();
-    const res = await request(app)
-      .post('/api/user/10000001/notify-changes')
-      .send({ fields: ['shyCoins', 'luckScore'] });
-    expect(res.status).toBe(200);
-    expect(res.body.notified).toBe(false);
-    expect(res.body.reason).toMatch(/no notifiable/i);
-  });
-
-  it('sends system PM for notifiable fields and returns notified=true', async () => {
-    const { sendSystemPm } = require('../../src/utils/system-pm');
-
-    const app = createAdminApp();
-    const res = await request(app)
-      .post('/api/user/10000001/notify-changes')
-      .send({ fields: ['displayName', 'profilePhotoUrl', 'shyCoins'] });
-    expect(res.status).toBe(200);
-    expect(res.body.notified).toBe(true);
-    expect(res.body.fields).toEqual(['displayName', 'profilePhotoUrl']);
-    expect(sendSystemPm).toHaveBeenCalledWith('10000001', expect.stringContaining('display name'));
-  });
-
-  it('returns 500 on error', async () => {
-    const { sendSystemPm } = require('../../src/utils/system-pm');
-    sendSystemPm.mockRejectedValueOnce(new Error('PM failed'));
-
-    const app = createAdminApp();
-    const res = await request(app)
-      .post('/api/user/10000001/notify-changes')
-      .send({ fields: ['displayName'] });
-    expect(res.status).toBe(500);
-    expect(res.body.error).toMatch(/internal server error/i);
-  });
-});
-
 // ─── POST /api/user/:uniqueId/warn — additional branches ────────────
-
-describe('POST /api/user/:uniqueId/warn — additional branches', () => {
-  it('returns 500 on unexpected createWarning error', async () => {
-    // Make the user doc fetch succeed, but the set (warning creation) fail
-    mockDocGet.mockResolvedValueOnce({
-      exists: true,
-      id: '10000001',
-      data: () => ({ gcsScore: 90, warningCount: 1 }),
-    });
-    getDoc.mockResolvedValueOnce({ displayName: 'Admin User' });
-    mockDocSet.mockRejectedValueOnce(new Error('Batch write failed'));
-
-    const app = createAdminApp();
-    const res = await request(app)
-      .post('/api/user/10000001/warn')
-      .send({ reason: 'Test error', severity: 2 });
-    expect(res.status).toBe(500);
-    expect(res.body.error).toMatch(/internal server error/i);
-  });
-
-  it('defaults severity to 3 when not provided', async () => {
-    mockDocGet.mockResolvedValueOnce({
-      exists: true,
-      id: '10000001',
-      data: () => ({ gcsScore: 100, warningCount: 0 }),
-    });
-    getDoc.mockResolvedValueOnce({ displayName: 'Admin User' });
-
-    const app = createAdminApp();
-    const res = await request(app)
-      .post('/api/user/10000001/warn')
-      .send({ reason: 'Default severity test' });
-    expect(res.status).toBe(200);
-    // severity 3 -> deduction 15
-    expect(res.body.deduction).toBe(15);
-    expect(res.body.newGcs).toBe(85);
-  });
-});
 
 // ─── GET /api/user/:uniqueId/warnings — startAfter branch ──────────
 
-describe('GET /api/user/:uniqueId/warnings — pagination', () => {
-  it('passes startAfter query param for pagination', async () => {
-    mockCollectionGet = jest.fn().mockResolvedValue({
-      docs: [
-        {
-          id: 'warn-5',
-          data: () => ({ reason: 'Older', severity: 1, createdAt: 1709913400000 }),
-        },
-      ],
-    });
-
-    const app = createAdminApp();
-    const res = await request(app).get(
-      '/api/user/10000001/warnings?startAfter=1709913500000&limit=5',
-    );
-    expect(res.status).toBe(200);
-    expect(res.body.warnings).toHaveLength(1);
-    expect(res.body.hasMore).toBe(false);
-  });
-
-  it('returns 500 when Firestore throws during list', async () => {
-    mockCollectionGet = jest.fn().mockRejectedValue(new Error('Query failed'));
-
-    const app = createAdminApp();
-    const res = await request(app).get('/api/user/10000001/warnings');
-    expect(res.status).toBe(500);
-    expect(res.body.error).toMatch(/internal server error/i);
-  });
-});
-
 // ─── POST /api/user/:uniqueId/warnings/:id/revoke — error branches ─
-
-describe('POST /api/user/:uniqueId/warnings/:warningId/revoke — additional', () => {
-  it('returns 404 when user not found during revoke', async () => {
-    getDoc.mockResolvedValueOnce({
-      id: 'warn-1',
-      revoked: false,
-      gcsDeduction: 10,
-    });
-    mockDocGet.mockResolvedValueOnce({ exists: false });
-
-    const app = createAdminApp();
-    const res = await request(app).post('/api/user/10000001/warnings/warn-1/revoke');
-    expect(res.status).toBe(404);
-    expect(res.body.error).toMatch(/user not found/i);
-  });
-
-  it('returns 500 on Firestore error during revoke', async () => {
-    getDoc.mockRejectedValueOnce(new Error('Firestore read error'));
-
-    const app = createAdminApp();
-    const res = await request(app).post('/api/user/10000001/warnings/warn-1/revoke');
-    expect(res.status).toBe(500);
-    expect(res.body.error).toMatch(/internal server error/i);
-  });
-});
 
 // ─── POST /api/user/:uniqueId/reset-gcs — error branch ─────────────
 
 // ─── GET /api/user/:uniqueId/stalkers — error branch ────────────────
-
-describe('GET /api/user/:uniqueId/stalkers — error handling', () => {
-  it('returns 500 when Firestore throws', async () => {
-    mockCollectionGet = jest.fn().mockRejectedValue(new Error('Stalkers query failed'));
-
-    const app = createAdminApp();
-    const res = await request(app).get('/api/user/10000001/stalkers');
-    expect(res.status).toBe(500);
-    expect(res.body.error).toMatch(/internal server error/i);
-  });
-});
 
 // ─── GET /api/conversations/:id/messages ────────────────────────────
 
@@ -1086,103 +583,8 @@ describe('POST /api/report-locks/:uniqueId/lock', () => {
 
 // ─── GET /api/user/:uniqueId/auth-status ────────────────────────────
 
-describe('GET /api/user/:uniqueId/auth-status', () => {
-  it('returns 403 for non-admin', async () => {
-    blockAdmin();
-    const app = createAdminApp();
-    const res = await request(app).get('/api/user/10000001/auth-status');
-    expect(res.status).toBe(403);
-  });
-
-  it('returns 404 when user not found', async () => {
-    getDoc.mockResolvedValueOnce(null);
-
-    const app = createAdminApp();
-    const res = await request(app).get('/api/user/10000001/auth-status');
-    expect(res.status).toBe(404);
-    expect(res.body.error).toMatch(/user not found/i);
-  });
-
-  it('returns 200 with auth status including biometric keys', async () => {
-    getDoc.mockResolvedValueOnce({
-      pinHash: 'hashed-pin',
-      pinSetAt: 1700000000000,
-      pinAttempts: 2,
-      pinLockedUntil: null,
-      pinLockoutCount: 1,
-    });
-    mockCollectionGet = jest.fn().mockResolvedValue({
-      docs: [
-        {
-          id: '10000001:device-abc',
-          data: () => ({ createdAt: 1700000000000 }),
-        },
-      ],
-    });
-
-    const app = createAdminApp();
-    const res = await request(app).get('/api/user/10000001/auth-status');
-    expect(res.status).toBe(200);
-    expect(res.body.pinSet).toBe(true);
-    expect(res.body.pinAttempts).toBe(2);
-    expect(res.body.isLocked).toBe(false);
-    expect(res.body.biometricKeys).toHaveLength(1);
-    expect(res.body.biometricKeys[0].deviceId).toBe('device-abc');
-  });
-
-  it('returns 500 on error', async () => {
-    getDoc.mockRejectedValueOnce(new Error('Read failed'));
-
-    const app = createAdminApp();
-    const res = await request(app).get('/api/user/10000001/auth-status');
-    expect(res.status).toBe(500);
-    expect(res.body.error).toMatch(/failed to get auth status/i);
-  });
-});
-
 // ─── POST /api/user/:uniqueId/reset-pin-lockout ────────────────────
 
 // ─── DELETE /api/user/:uniqueId/biometric-keys/:deviceId ────────────
 
 // ─── GET /api/metrics/otp ───────────────────────────────────────────
-
-describe('GET /api/metrics/otp', () => {
-  it('returns 403 for non-admin', async () => {
-    blockAdmin();
-    const app = createAdminApp();
-    const res = await request(app).get('/api/metrics/otp');
-    expect(res.status).toBe(403);
-  });
-
-  it('returns default values when metrics doc does not exist', async () => {
-    mockDocGet.mockResolvedValueOnce({ exists: false });
-
-    const app = createAdminApp();
-    const res = await request(app).get('/api/metrics/otp');
-    expect(res.status).toBe(200);
-    expect(res.body).toEqual({ count: 0, date: null, limit: 100 });
-  });
-
-  it('returns metrics when doc exists', async () => {
-    mockDocGet.mockResolvedValueOnce({
-      exists: true,
-      data: () => ({ count: 42, date: '2026-04-01' }),
-    });
-
-    const app = createAdminApp();
-    const res = await request(app).get('/api/metrics/otp');
-    expect(res.status).toBe(200);
-    expect(res.body.count).toBe(42);
-    expect(res.body.date).toBe('2026-04-01');
-    expect(res.body.limit).toBe(100);
-  });
-
-  it('returns 500 on error', async () => {
-    mockDocGet.mockRejectedValueOnce(new Error('Firestore read failed'));
-
-    const app = createAdminApp();
-    const res = await request(app).get('/api/metrics/otp');
-    expect(res.status).toBe(500);
-    expect(res.body.error).toMatch(/failed to get otp metrics/i);
-  });
-});
